[WIP][scripts] Restore check_cipher.sh
As mentioned in 74b6285, this script did nothing because it waited for
scenario   to   be   0,  while   collect_cipher_metrics.sh   forwarded
scenario=1.

Now check_cipher.sh will always compile  scenario 0 and check that all
"CORRECT!" outputs are shown.

Remove _check_ciphers  in felics-run, since  this is now  performed by
check_cipher.sh.

Note that AVR  outputs are broken right now,  so check_cipher.sh fails
for this platform:

    O:Plaintext:
    ^[[32mPlaintext:..
    ^[[0mO:00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
    ^[[32m00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 ..

NB: left indentation as-is to make the diff easier to read.