Move linear layer to assembly
Lilliput-II-128 (felicsref, -O3): 6162 592 149366
Lilliput-II-128 (msp, -O3):       4700 530 138874

We lose a  few cycles (136k ↗  138k).  AFAWCT this is  due to function
call overhead, since the assembly is  identical to the one produced by
the reference version.

We'll try to do both the nonlinear  and linear layers in the same loop
to win back some cycles.