Update ΘCB3 reference code to match paclido/sp3
Fix bug  in tweak construction  introduced in e5838b7;  optimize nonce
handling.

Performance changes:

On AVR:
    ROM: -11%
    RAM: +0.3% (1 byte)

On MSP:
    ROM: -17% (felicsref), -25% (msp)

On ARM:
    ROM: -15%

Marginal  (<1%) improvements  to  execution speed.   This is  expected
since  the  test vector  is  only  one-block-long: the  nonce-handling
optimization  ensures that  the nonce  is copied  into the  tweak once
instead of l times, so there should be no speedup when l=1.