[msp-smallrom] Add new implementation
Also add  field in implementation.info  to limit an  implementation to
some compilation flag, so that when we run e.g.

    felics-run Lilliput-*_vfelicsref Lilliput-*_vmsp-fast       \
               Lilliput-*_vmsp-smallrom --options="-O3;-Os"

"msp-smallrom" is only run with -Os.

I am  not sure I  like this  solution; ideally I  would like a  way to
specify  arbitrary  limitations on  (cipher,  cflags)  pairs from  the
command-line,    but    as    long    as    felics-run    relies    on
collect_ciphers_metrics.sh this will not be easy to implement.