Fix cold-cache issue with RAM measurement on PC
My machine has a "spinning rust" disk; cipher_ram.sh would
consistently fail
- the first time it was run,
- right after dropping caches (echo 3 > /proc/sys/vm/drop_caches).

Instead of running "sleep 1" and hoping for the best, run GDB in the
foreground.

The -tty switch is needed to prevent a weird interaction with the
timeout(1) command[1].  Another way to side-step the issue would be to
give --foreground to timeout(1); however "in this mode, children of
COMMAND will not be timed out".  I don't know if that's a problem
(empirically it seems not); I figure fixing this locally is better
than risking side-effects.

[1]: https://notmuchmail.org/pipermail/notmuch/2019/028017.html