Remove obsolete package for Docker installation
Cf. previous commit.  We could replace cpufrequtils with
linux-tools-common, but the guest distro's cpupower might not be
compatible with the eventual host's kernel.

Not sure what the best approach is; right now at least the user will
get a warning at runtime will set them on the path to
fixup-cpupower.sh, which will install the correct version of cpupower.